Common Component Button and font increase
