+++ client/modules/_common_components/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/common-components';
 export { Spinner } from './lib/Spinner';
 export { Icon } from './lib/Icon';
+export { SecondaryButton, PrimaryButton } from './lib/Button';
