+++ client/modules/_common_components/src/lib/Button/index.ts
@@ -1,0 +1,1 @@
+export * from './Button';
